nvme: host: Introduce nvme_submit_async_cmd()
Introduce the functions __nvme_submit_async_cmd() and
nvme_submit_async_cmd() to allow executing commands asynchronously with
the command completion handled through a callback function that can be
optionally provided by the command issuer. The structure
nvme_async_cmd_endio_cb is added to define a command completion callback
and the data to pass to that callback.

Signed-off-by: Damien Le Moal <dlemoal@kernel.org>