arm64: dts: rk3588-rock-5b: enable SATA for M.2 E-Key to rock-5b
We need to disable pcie2x1l0 in order to enable the onboard SATA
controller, as the top M.2 E-Key slot can be used by either the onboard
PCIe controller or the onboard SATA controller, but not both.

The M.2 E-Key slot is connected to a rockchip USB/PCIe/SATA combo PHY.
The SATA device tree node has a phandle to the combophy and supplies
the PHY mode which the PHY should be configured in (PHY_TYPE_SATA).

Signed-off-by: Niklas Cassel <niklas.cassel@wdc.com>