PCI: dwc: endpoint: fix resizable BAR capability handling
Commit fc9a77040b04 ("PCI: designware-ep: Configure Resizable BAR cap to
advertise the smallest size") added code that claims it modifies the
Resizable BAR capability to only advertize support for 1 MB size BARs.

There are however two problems with that commit:
1) It writes all zeroes to PCI_REBAR_CAP (which advertizes possible
BAR sizes that a BAR be resized to). However, according to the spec,
it is not legal to have no bit set in PCI_REBAR_CAP.
If you want to advertize support for a 1 MB BAR size, which is the
smallest allowed by the spec, bit 4 should be set.

2) To actually control which BAR size that is currently in use,
you should set REBAR_CTRL_REG_BAR_SIZE field in the PCI_REBAR_CTRL reg.
To select the smallest BAR size (1 MB), we should write all zeroes to
this field. This wasn't done before, which could mean that a BAR size that
wasn't even supported was the one currently in use. Since some bits are
read-only in this reg, just write zeroes to the whole reg.

Fixes: fc9a77040b04 ("PCI: designware-ep: Configure Resizable BAR cap to advertise the smallest size")
Signed-off-by: Niklas Cassel <cassel@kernel.org>