PCI: rockchip-ep: Improve link training
The Rockchip rk339 technical reference manual describe the endpoing mode
link training process clearly and states that:
  Insure link training completion and success by observing link_st field
  in PCIe Client BASIC_STATUS1 register change to 2'b11. If both side
  support PCIe Gne2 speed, re-train can be Initiated by asserting the
  Retrain Link field in Link Control and Status Register. The software
  should insure the BASIC_STATUS0[negotiated_speed] change to "1", that
  indicates re-train to Gen2 successfully.
This procedure is very similar to what is done for the root-port mode in
rockchip_pcie_host_init_port().

Implement this link training procedure for for the endpoint mode as
well. Given that the rk3399 SoC does not have an interrupt signaling
link status changes, training is implemented as a delayed work which is
rescheduled until the link training completes of the endpoint is
stopped. The link training work is first scheduled in
rockchip_pcie_ep_start() when the endpoint function is started.
Link ctraining completion is signaled to the function using
pci_epc_linkup(). Accordingly, the linkup_notifier field of the rockchip
pci_epc_features structure is changed to true.

Signed-off-by: Damien Le Moal <dlemoal@kernel.org>