lib/group_cpus: drop unneeded cpumask_empty() call in __group_cpus_evenly()
The function is called twice.  First time it's called with cpumask_present
as a parameter, which can't be empty.  Second time it's called with a mask
created with cpumask_andnot(), which returns false if the result is an
empty mask.

We can safely drop redundant cpumask_empty() call from the
__group_cpus_evenly() and save few cycles.

Link: https://lkml.kernel.org/r/20231228200936.2475595-7-yury.norov@gmail.com
Signed-off-by: Yury Norov <yury.norov@gmail.com>
Reviewed-by: Ming Lei <ming.lei@redhat.com>
Cc: Andy Shevchenko <andriy.shevchenko@linux.intel.com>
Cc: Rasmus Villemoes <linux@rasmusvillemoes.dk>
Cc: Thomas Gleixner <tglx@linutronix.de>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>