block: Introduce zone write plugging
Introduce struct blk_zone_wplug to implement zone write plugging to
guarantee that for zoned block devices, at any time, at most one write
per zone is in flight. This mechanism replaces zone write locking
which is controlled at the scheduler level but only implemented by
mq-deadline.

Zone write plugging operates at a higher level in the submission path,
on BIOs instead of requests. A zone write plug is simply a BIO list that
is atomically manipulated using a lock implemented with a bit (a flag
of the zone write plug) and a kblockd submission work. A write BIO to a
zone of a zoned block device is "plugged" to delay its execution if
another write BIO was already issued to the same zone and has not yet
completed or if other BIOs are already waiting execution in the BIO
list of the zone plug.

Plugging of zone write BIOs is done using the function
blk_zone_write_plug_bio() which returns false if a BIO execution does
not need to be delayed and true otherwise. This function is called
from blk_mq_submit_bio() after a BIO is split to avoid large BIOs
spanning multiple zones which would cause mishandling of zone write
plugging. This enables by default zone write plugging for any mq
request-based block device.

Zone write plugging can also be used by BIO-based drivers by modifying
the driver to expliclty call blk_zone_write_plug_bio() in its
->submit_bio method. For such device, blk_zone_write_plug_bio() handles
splitting of alrge BIOs straddling zone boundaries.

Only write and write zeroes BIOs are plugged. This modification does not
introduce any overhead for other operations. A BIO that is being
handled through zone write plugging is flagged using the new BIO flag
BIO_ZONE_WRITE_PLUGGING. BIOs that are flagged are processed one by one,
with every completion of a flagged BIO triggering the execution of the
next BIO in the zone write plug. The function blk_zone_write_bio_endio()
is called from bio_endio() to schedule the zone write plug bio work.
This work "unplugs" the first BIO in the plug BIO list and issues it
using submit_bio_noacct_nocheck(). This mechanism allows multiple zones
to be written simulatenously.

Zone write plugging ignores flush BIOs without data. Hovever, any flush
BIO that has data is always plugged so that the write part of the flush
sequence is serialized with other regular writes.

To match the new data structures used for zoned disks, the function
disk_free_zone_bitmaps() is renamed to the more generic
disk_free_zone_resources().

Given that any BIO handled through zone write plugging will be the only
BIO in flight for the target zone when it is executed, there will be no
chances of merging newly incoming BIOs with the BIO in flight. All BIOs
handled through zone write plugging are thus marked with REQ_NOMERGE to
not attempt merging.

For the rare case where the queue limits change while a zone write BIO
is plugged, causing the BIO to be split again when it is unplugged and
submitted, the function blk_zone_write_plug_split_bio() is used in
__bio_split_to_limits() to re-plug the second half of the split BIO and
to mark the first half with the BIO_ZONE_WRITE_PLUGGING flag and allow
it to proceed with execution similarly as if no split had happen.

Signed-off-by: Damien Le Moal <dlemoal@kernel.org>