block: Introduce zone write plugging
Introduce struct blk_zone_wplug to implement zone write plugging to
guarantee that for zoned block devices, at any time, at most one write
per zone is in flight. This mechanism replaces zone write locking
which is controlled at the scheduler level and implemented only by
mq-deadline.

Zone write plugging operates at a higher level in the submission path,
on BIOs instead of requests. A zone write plug is simply a BIO list that
is atomically manipulated using a bit spinlock (flag of the zone write
plug) and a kblockd submission work. A write BIO to a zone of a zoned
block device is "plugged" to delay its execution if another write BIO
was already issued to the same zone and has not yet completed or if
other BIOs are already waiting for execution in the BIO list of the zone
plug.

This mechanism allows to:
 - Remove the restriction on using only mq-deadline for zoned block
   devices. Any block IO scheduler, including "none" can be used.
 - Zone write plugging operates on BIOs instead of requests. Plugged
   BIOs waiting for execution thus do not hold scheduling tags and thus
   are not preventing other I/Os to porceed (reads or writes to other
   zones). This significantly improve the device use and overall
   performance.
 - The code is less invasive in the block layer and is mostly limited to
   blk-zoned.c with some small changes in iblk-mq.c, blk-merge.c and
   bio.c.

One drawback of this mechanism for now is that it prevents merging of
small sequential writes to a zone. Performance tests showed that this
is not an issue for fast devices (e.g. NVMe ZNS), nor for SMR HDDs using
a write cache. Merging of sequential plugged BIOs can be added latter
when constructing a request for a BIO that was unplugged. This commit
does not address this.

Plugging of zone write BIOs is done using the function
blk_zone_write_plug_bio() which returns false if a BIO execution does
not need to be delayed and true otherwise. This function is called
from blk_mq_submit_bio() after a BIO is split to avoid large BIOs
spanning multiple zones which would cause mishandling of zone write
plugging. This enables by default zone write plugging for any mq
request-based block device.

Zone write plugging can also be used by BIO-based drivers by modifying
the driver to expliclty call blk_zone_write_plug_bio() in its
->submit_bio method. For such device, the driver must ensure that a
BIO passed to blk_zone_write_plug_bio() is already split and not
straddling zone boundaries.

Only write and write zeroes BIOs are plugged. This modification does not
introduce any overhead for other operations. A BIO that is being
handled through zone write plugging is flagged using the new BIO flag
BIO_ZONE_WRITE_PLUGGING. BIOs that are flagged are processed one by one,
with every completion of a flagged BIO triggering the execution of the
next BIO in the zone write plug. The function blk_zone_write_bio_endio()
is called from bio_endio() to schedule the zone write plug bio work.
This work "unplugs" the first BIO in the plug BIO list and issues it
using submit_bio_noacct_nocheck(). This mechanism allows multiple zones
to be written simulatenously.

Zone write plugging ignores flush BIOs without data. Hovever, any flush
BIO that has data is always plugged so that the write part of the flush
sequence is serialized with other regular writes.

To match the new data structures used for zoned disks, the function
disk_free_zone_bitmaps() is renamed to the more generic
disk_free_zone_resources().

Given that any BIO handled through zone write plugging will be the only
BIO in flight for the target zone when it is executed, there will be no
chance of merging newly incoming BIOs with requests (plugged in the
current context or inserted in the scheduler). All BIOs handled through
zone write plugging are thus marked with REQ_NOMERGE to not attempt
merging.

For the rare case where the queue limits change while a zone write BIO
is plugged, causing the BIO to be split again when it is unplugged and
submitted, the function blk_zone_write_plug_split_bio() is used in
__bio_split_to_limits() to re-plug the second half of the split BIO and
to mark the first half with the BIO_ZONE_WRITE_PLUGGING flag, thus
allowing it to proceed with execution as if no split had happen.

When BIOs are plugged in a zone write plug, the device request queue
usage counter is always incremented. This reference is dropped
when the BIO unplugged, submitted and completes, with
blk_mq_submit_bio() modified to ensure that another usage counter
reference is not taken for BIOs that have gone through zone write
plugging. This ensures that all plugged BIOs can be executed even if a
different context requested a queue freeze: as plugged BIOs hold an
extra reference, blk_freeze_queue() will not return until all plugged
BIOs are first processed.

This commit contains contributions from Christoph Hellwig <hch@lst.de>.

Signed-off-by: Damien Le Moal <dlemoal@kernel.org>