block: Reduce zone write plugging memory usage
With zone write plugging, each zone of a zoned block device has a
64B struct blk_zone_wplug. While this is not a problem for small
capacity drives with few zones, this structure size result in large
memory usage per device for large capacity disks. E.g., for a 28 TB SMR
disk with over 104,000 zones of 256 MB, the zone write plug array uses
6.6 MB of memory.

However, except for the zone write plug flags, zone capacity and zone
write pointer offset that need to be always available (the later 2 to
avoid having to do too many report zones), the remaining fields of
struct blk_zone_wplug are needed only when a zone is being written to.

This commit introduces struct blk_zone_active_wplug to reduce the size
of struct blk_zone_wplug from 64B down to 16B. This is done using an
union of a pointer to a struct blk_zone_active_wplug and of the zone
write pointer offset and zone capacity, with the zone write plug flags
left as the first field of struct blk_zone_wplug.

The flag BLK_ZONE_WPLUG_ACTIVE is introduced to indicate if the pointer
to struct blk_zone_active_wplug of a zone write plug is valid. For such
case, the write pointer offset and zone capacity fields are accessible
from struct blk_zone_active_wplug. Otherwise, they can be accessed from
struct blk_zone_wplug.

This data structure organization allows tracking the write pointer
offset of zones regardless of the zone write state (active or not).
Handling of zone reset, reset all and finish operations are modified
to update a zone write pointer offset according to its state.

A zone is activated in __blk_zone_lookup_wplug() with a call to the new
function blk_zone_activate_wplug(). Reclaiming of the active zone write
plugs allocated is done after a zone becomes full or is reset and
becomes empty. Reclaiming (freeing) of a zone active write plug
structure is done either directly when a plugged BIO completes and the
zone is marked as full (BLK_ZONE_WPLUG_FULL flag), or when resetting or
finishing zones.

With this mechanism, the amount of memory used per block device for zone
write plugs is reduced by a factor of 4 (when there are no active zone
write plugs). E.g. for a 28 TB SMR disk, memory usage is reduce to 1.6
MB.

This commit contains contributions from Christoph Hellwig <hch@lst.de>.

Signed-off-by: Damien Le Moal <dlemoal@kernel.org>