block: Introduce bio_rw_needs_split()
Introduce the function bio_rw_needs_split() to check if a read/write
BIO needs splitting and its number of segments. This function is
implemented by factoring out the first part of bio_split_rw() into the
helper function __bio_rw_needs_split(). bio_split_rw() is modified to
use this helper to determin if bio_split() needs to be called.

While at it, remove the discard mention in bio_split_rw() kdoc comment
as this function is not used to handle discard BIO split.

Signed-off-by: Damien Le Moal <dlemoal@kernel.org>