block: Use a mempool to allocate zone write plugs
Allocating zone write plugs using a struct kmem_cache does not guarantee
that enough plugs can be allocated to simultaneously write up to the
maximum number of active zones of a zoned block device.

Improve this by using a mempool with a size equal to the disk maximum
number of open zones or maximum number of active zones, whichever is
larger. For zoned devices that do not have open or active zone limits,
the default 128 is used as the mempool size. If a change to the zone
limits is detected, the mempool is resized in
blk_revalidate_disk_zones().

Signed-off-by: Damien Le Moal <dlemoal@kernel.org>