nvme: Add PCI transport type
Define the transport type NVMF_TRTYPE_PCI for PCI endpoint targets.
This transport type is defined as transport "0" and thus will be the
default for a port and cannot be used as the value for the TRTYPE field
of a Discovery Log Page entry (because value "0" is reserved).

The helper functions nvmet_is_pci_ctrl() is also introduced to help
identify a target controller using the PCI transport.

Signed-off-by: Damien Le Moal <dlemoal@kernel.org>