nvmet: New NVMe PCI endpoint target driver
Implement a PCI target driver using the PCI endpoint framework. This
requires a PCI controller capable of running in endpoint mode.

The PCI endpoint framework is used to set up a PCI endpoint and its BAR
compatible with a NVMe PCI controller. The framework is also used to map
local memory to the PCI address space to execute MMIO accesses for
retreiving NVMe commands from submission queues and posting completion
entries to completion queues. If supported, DMA is used for command data
transfers as well as PRP lists or SGL segments retrieval.

The NVMe target driver relies on the NVMe target core to execute all
commands isssued by the host. Most of the PCI target driver code does:
 - Initialization and teardown of the endpoint controller and its
   backend target PCI controller.
 - Manage the endpoint controller state according to actions from the
   host (e.g. CC.EN register) and propagate these actions to the PCI
   target controller.
 - Manage the controller submission and completion queues using local
   memory. For submission queues, this management includes a polling
   mechanism of the submission queues doorbells to detect the submission
   of new commands by the host.
 - Manage the execution of admin and I/O commands using the PCI target
   controller. The endpoint controller is only in charge of parsing a
   command PRPs or SGLs to setup a command local memory data buffer and
   its set of PCI address segments for data transfer. The execution of
   commands is entirely done using nvmet requests handled by the PCI
   target controller.
 - Handle data transfers through mmio or DMA before executing a command
   (for write commands) or after the PCI target controller signals a
   command completion (for read commands).

The configuration of a NVMe PCI endpoint controller is done using
configfgs. First the NVMe PCI target controller configuration must be
done setting up a subsystem and a port with the "pci" addr_trtype
attribute. The susbsystem can be setup using a file or block device
backed namespace or a passthrough NVMe device. After this, the PCI
endpoint can be configured and bound to the PCI endpoint controller to
start the NVMe endpoint controller.

Using a Rock5B board (Rockchip RK3588 SoC) with a null_blk block device
as a namespace, the maximum performance achieved is 44 KIOPS (4K random
reads) and up to 2.4 GB/S throughput.

Signed-off-by: Damien Le Moal <dlemoal@kernel.org>