nvmet: Improve nvmet_alloc_ctrl() interface and implementation
Introduce struct nvmet_alloc_ctrl_args to define the arguments for
the function nvmet_alloc_ctrl() to avoid the need for passing a pointer
to a struct nvmet_req. This new data structure aggregates together the
arguments that were passed to nvmet_alloc_ctrl (subsysnqn, hostnqn and
kato), together with the struct nvmet_req fields used by
nvmet_alloc_ctrl(), that is: port, p2p_client, and ops fields as input
and the result and error_loc fields as output, as well as a status
field. nvmet_alloc_ctrl() is also changed to return a pointer to the
allocated controller instead of a status code, as the status is now
returned through the status field of struct nvmet_alloc_ctrl_args.

The function nvmet_setup_p2p_ns_map() is also changed to not take a
pointer to a struct nvmet_req as argument, instead, directly specify
the p2p_client device pointer needed as argument.

The in nvmet_execute_admin_connect() that initializes a new target
controller after allocating it is moved into nvmet_alloc_ctrl().
The code that sets up an admin queue for the controller (and the call
to nvmet_install_queue()) remains in nvmet_execute_admin_connect().

Finally, export nvmet_alloc_ctrl() to allow target drivers to use this
function directly to allocate an initialize a new controller structure
without the need to rely on a fabrics connect command.

Signed-off-by: Damien Le Moal <dlemoal@kernel.org>