nvmet: Add get_vid() fabrics operation
Allow a target controller to specify a vendor ID (and subsystem vendor
ID) with the new fabrics operation get_vid(). This operation is used
from nvmet_execute_identify_ctrl() to fill the vid and ssvid fields of
the indetify controller data structure.

This new operation is optional. If not defined, the vid and ssvid fields
are zero filled (as they were before this change).

Signed-off-by: Damien Le Moal <dlemoal@kernel.org>